requirements.txt
numpy==1.19.5
onnxruntime==1.8.1
Pillow==8.1.2
tensorflow==2.5.0
